Renamed coin_shuffle.py to the more appropriate round.py
It irked me that the file was called "coin_shuffle.py" (an
inconveniently long name) where the only thing living in it is a (well
named) class 'Round'.

Fixed.
